Allow LITeleporters to be enabled/disabled
Added a variable _enabled to LITeleporter that controls whether the
teleporter is enabled or not. Enabling/disabling teleporters can be
done using the triggers "enable", "disable" and "toggle".
The toggle trigger enables the teleporter if it is currently disabled,
and disables it if is currently enabled. Since toggling component can
be used for many types of components in the future, I decided to
implement it generically. Whether the teleporter starts
enabled is determined by the property isEnabled in LIProperties.